+++ tabbycat/motions/urls_admin.py
@@ -17,7 +17,11 @@ urlpatterns = [
             views.UnreleaseMotionsView.as_view(),
             name='motions-unrelease'),
         path('display/',
-            views.DisplayMotionsView.as_view(),
+            views.AdminDisplayMotionsView.as_view(),
             name='motions-display'),
     ])),
+
+    path('statistics/',
+        views.MotionStatisticsView.as_view(),
+        name='motions-statistics'),
 ]