+++ tabbycat/motions/urls_assistant.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('display/',
+        views.AssistantDisplayMotionsView.as_view(),
+        name='motions-assistant-display'),
+]