+++ tabbycat/motions/urls_public.py
@@ -4,4 +4,5 @@ from . import views
 
 urlpatterns = [
     path('', views.PublicMotionsView.as_view(), name='motions-public'),
+    path('statistics/', views.PublicMotionStatisticsView.as_view(), name='motions-public-statistics'),
 ]